requirements.txt
Werkzeug==1.0.1
Flask==1.1.2
numpy==1.18.5
tensorflow_gpu==2.3.1
Pillow==8.0.0
tensorflow==2.3.1
